requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy==1.22.3
opencv-python==4.5.5.64
torch==1.10.0
torchvision==0.11.1
Pillow==9.0.1
requests==2.27.1
python-multipart==0.0.9